# Turn off ErrorBuilders during LSP queries