… core/GlobalState.cc
… test/testdata/lsp/completion/constants_all_kinds.rb