<!-- (optional) Explain your change, focusing on the details of the solution. This is a great place to call out user-visible changes. -->


### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

We drop any errors discovered when making a query on the floor, so it
doesn't make sense to spend time in resolver making those error messages
better by adding "Did you mean" suggestions for missing constants.

In particular, the findMemberFuzzyMatch search is crazy slow.

As measured on //test:test_LSPTests/testdata/disabled/lsp/completion/constants_all_kinds
in debug mode, this change improves perf when testing on my MacBook Pro:

- from 7s to 2.5s (when ENFORCEs don't have Timers)
- from 42s to 9s (when ENFORCEs have Timers)

I'm still not sure why adding ENFORCEs to Timers makes it so much
slower, but I plan to timebox looking into that too.

Regardless, this should have a huge effect on our entire test suite.


### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

See included automated tests.