@jez 2020-04-06
cc @jvilk-stripe I expect this change to speed up certain kinds of LSP queries.