@jvilk-stripe 2020-04-07
I'm fine with this! At some point we'll probably want to make `findMemberFuzzyMatch` faster.